using combinedCodeFix unusedIdentifier_delete
