+++ src/compiler/tools/index.ts
@@ -1,6 +1,5 @@
 export * from "./inputs";
 export * from "./LanguageService";
 export * from "./Program";
-export * from "./refactors";
 export * from "./results";
 export * from "./TypeChecker";
