--- src/compiler/tools/refactors/index.ts
@@ -1,1 +1,0 @@
-export * from "./removeUnusedDeclarations";
